rename http.py to client.py due to causing No Module Found error for http.cookiejar
